Add Hawkeye cache replacement policy (Jain & Lin, ISCA 2016)
OPTgen reconstructs Belady's optimal decisions for the observed PC/address stream; a PC-indexed 8K×3-bit saturating-counter predictor labels each PC as cache-friendly (≥4) or cache-averse (<4). Friendly lines install at RRPV=0, averse at RRPV=7; demand hits decrement RRPV; victim selection is SRRIP-style. IReplacementPolicy gains two new default-body hooks (SetPendingAddress, RecordHitPc) so SetAssociativeCache can feed the policy on both misses and hits without touching other policies.

+++ Mechanism/BranchPredictModels/LlbpBranchPrediction.cs
@@ -76,7 +76,7 @@ public class LlbpPredictor : TageScLPredictor {
         }
         else if (provPred != taken) {
             int allocTable = _lastProvider + 1;
-            if ((uint)allocTable < (uint)LTagePredictor.NumTables)
+            if ((uint)allocTable < LTagePredictor.NumTables)
                 _storage.GetOrCreate(_llbpCtxKey).AllocateIfAbsent(PatternKey(pc, allocTable), taken);
         }
     }
